fixed empty file management

+++ script.sh
@@ -10,22 +10,18 @@ if [ -z $1 ]
 then
 	echo "usage: ./script.sh [option]"
 else
-	if test -f $ILINK
-	then
-		rm -f $ILINK
-	fi
-	if test -f $OLINK
-	then
-		rm -f $OLINK
-	fi
 	./generator $1 > $IPATH
-	./lem-in -i $IPATH -o $OPATH
-	ln -s $IPATH $ILINK
-	ln -s $OPATH $OLINK
-	a=$(wc -l $IPATH | awk '{printf "%s\n", $1}')
-	b=$(wc -l $OPATH | awk '{print $1}')
-	OUR=$(expr $b - $a - 1)
-	PERF=$(sed '2q;d' $IPATH | tr -dc '0-9')
-	echo "Result: $OUR/$PERF"
-	python3 checker.py < $IPATH
+	if [ -s $IPATH ]
+	then
+		rm -f $ILINK $OLINK
+		./lem-in -i $IPATH -o $OPATH
+		ln -s $IPATH $ILINK
+		ln -s $OPATH $OLINK
+		a=$(wc -l $IPATH | awk '{printf "%s\n", $1}')
+		b=$(wc -l $OPATH | awk '{print $1}')
+		OUR=$(expr $b - $a - 1)
+		PERF=$(sed '2q;d' $IPATH | tr -dc '0-9')
+		echo "Result: $OUR/$PERF"
+		python3 checker.py < $IPATH
+	fi
 fi
